+++ userBack/src/main/java/com/fisioplus/user_backend/repository/ProgresoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProgresoRepository extends JpaRepository<Progreso, Long> {
-    List<Progreso> findByPacienteId(Long pacienteId);
+public interface ProgresoRepository extends JpaRepository<Progreso, Integer> {
+    List<Progreso> findByPacienteId(Integer pacienteId);
 }
