+++ userBack/src/main/java/com/fisioplus/user_backend/service/ProgresoService.java
@@ -5,6 +5,6 @@ import com.fisioplus.user_backend.dto.ProgresoDTO;
 import java.util.List;
 
 public interface ProgresoService {
-    List<ProgresoDTO> listarPorPaciente(Long pacienteId);
+    List<ProgresoDTO> listarPorPaciente(Integer pacienteId);
     ProgresoDTO guardar(ProgresoDTO dto);
 }
